adapters/odrive/pyproject.toml
[project]
name = "polyflow-odrive-adapter"
version = "0.1.0"
description = "Polyflow hardware adapter for ODrive S1 controllers (CAN and USB transports)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Polyflow Robotics" }]

dependencies = [
    "polyflow-adapter-sdk>=0.1.0",
    "python-can>=4.0",
    "odrive>=0.6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
